# packages/contracts/openapi/labelhub.yaml
openapi: 3.0.3
info:
  title: LabelHub API
  version: 0.10.0
  description: Contract-first API for the LabelHub modular monolith and AI worker.
servers:
  - url: http://localhost:8080/api
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
      responses:
        '200':
          description: Owner tasks for the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTasks'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
    post:
      tags: [Tasks]
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task detail visible to the authenticated Owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    patch:
      tags: [Tasks]
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: 永久删除 Owner 任务
      description: 永久删除该 task 以及所有 task 范围内的事实数据(sessions、submissions、ai_calls、quality_ledger_entries、verdicts、export_snapshots、task_transitions 等)。本操作覆盖 M6-P0.5 提出的 submission 不可变事实承诺,详细记录见 docs/internal/m6p7-verification.md 的 R8 段。
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task permanently deleted.
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /tasks/{taskId}/submissions:
    get:
      tags: [Tasks]
      operationId: listOwnerTaskSubmissions
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Owner-visible submission summaries for a task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedOwnerSubmissions'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /tasks/{taskId}/ai-prereview/summary:
    get:
      tags: [Tasks]
      operationId: getTaskAiPrereviewSummary
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Owner-visible AI prereview aggregate status for all eligible task submissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAiPrereviewSummary'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /tasks/{taskId}/ai-prereview/enqueue:
    post:
      tags: [Tasks]
      operationId: enqueueTaskAiPrereviews
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '202':
          description: Eligible task submissions were accepted into the asynchronous AI prereview queue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAiPrereviewEnqueueResult'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /tasks/{taskId}/workflow-progress:
    get:
      tags: [Tasks]
      operationId: getTaskWorkflowProgress
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Owner-visible aggregate progress for the task labeling and review workflow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskWorkflowProgress'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /tasks/{taskId}/transitions:
    get:
      tags: [Tasks]
      operationId: getTaskTransitions
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Append-only task transition history in chronological order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskTransition'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /tasks/marketplace:
    get:
      tags: [Tasks]
      operationId: listMarketplaceTasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: q
          in: query
          schema:
            type: string
          description: Search task title, description, and tags.
        - name: tag
          in: query
          schema:
            type: string
          description: Match a published task tag exactly.
        - name: hasReward
          in: query
          schema:
            type: boolean
          description: When true, only return tasks with a reward rule.
        - name: deadline
          in: query
          schema:
            type: string
            enum: [day, week]
          description: Limit to tasks whose deadline is within one day or one week.
      responses:
        '200':
          description: Available tasks for labelers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMarketplaceTasks'
  /tasks/{taskId}/transition:
    patch:
      tags: [Tasks]
      operationId: transitionTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTransitionRequest'
      responses:
        '200':
          description: Updated task state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
  /tasks/{taskId}/current-dataset:
    patch:
      tags: [Tasks]
      operationId: updateTaskCurrentDataset
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskCurrentDatasetRequest'
      responses:
        '200':
          description: Task current dataset updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
  /tasks/{taskId}/schema-from-template:
    post:
      tags: [Tasks]
      operationId: applySchemaTemplateToTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplySchemaTemplateRequest'
      responses:
        '201':
          description: Created a task-bound schema copy from a reusable template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplySchemaTemplateResult'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
...
  /tasks/{taskId}/claim:
    post:
      tags: [Tasks]
      operationId: claimTaskItem
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '201':
          description: Claimed annotation session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /tasks/{taskId}/claim-batch:
    post:
      tags: [Tasks]
      operationId: claimTaskItems
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimTaskItemsRequest'
      responses:
        '201':
          description: Claimed annotation sessions. Fewer sessions may be returned when fewer items are available.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimTaskItemsResult'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
  /tasks/{taskId}/submit-drafts:
    post:
      tags: [Tasks]
      operationId: submitTaskDrafts
      description: Submit all editable sessions claimed by the authenticated labeler for this task. The current session uses the provided payload; other sessions use their latest saved drafts.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitTaskDraftsRequest'
      responses:
        '201':
          description: Submissions created and queued for AI review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitTaskDraftsResult'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorBadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorUnauthorized:
      description: Missing, expired, or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorForbidden:
      description: Authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    TaskStatus:
      type: string
      enum: [draft, published, paused, ended]
...
    CreateTaskRequest:
      type: object
      required: [title, quotaTotal, deadlineAt]
      properties:
        title:
          type: string
        description:
          type: string
        instructionRichText:
          type: string
        tags:
          type: array
          items:
            type: string
        rewardRule:
          type: object
          additionalProperties: true
        deadlineAt:
          type: string
          format: date-time
        quotaTotal:
          type: integer
...
    PagedTasks:
      type: object
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
...